feat: bloc intro candidature spont plus refacto fichiers
